+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.2 branch
 ### 1.2.11
+* added `debug()` and `dd()` global function;
 * all `_or_fail()` functions can take a string or a `Throwable` instance as
     `$exception` parameter;
 * `FileNotExistsException`, `NotDirectoryException`, `NotReadableException` and